# ApparelPro.WebApi/APIModels/Reference/CreatePortDestinationAPIModel.cs
namespace ApparelPro.WebApi.APIModels.Reference
{
    public class CreatePortDestinationAPIModel
    {
        public int Id { get; set; }
        public string CountryCode { get; set; }
        public string DestinationName { get; set; }
    }
}