# ApparelPro.WebApi/APIModels/Reference/PortDestinationAPIModel.cs
namespace ApparelPro.WebApi.APIModels.Reference
{
    public class PortDestinationAPIModel
    {
        public int Id { get; set; }
        public string CountryCode { get; set; } = string.Empty;
        public string DestinationName { get; set; } = string.Empty;
    }
}